fix(socket): address Copilot review — Win64 handle type, narrowing guards, WSAStartup race, header rename
- socket.hpp: sock_type_t is now `typedef SOCKET sock_type_t` under _WIN32
  (was `unsigned int`, which truncates the pointer-sized SOCKET/UINT_PTR on
  64-bit Windows). INVALID_SOCKET sentinel usage lines up unchanged.
- socket.hpp: template set_option now static_asserts std::is_trivially_copyable_v<T>
  since it forwards raw object bytes to setsockopt(); added <type_traits>.
- socket.cpp: guard size_t->int narrowing in set_option (reject size > INT_MAX)
  and in set_receive_buffer_size / set_send_buffer_size (reject bytes > INT_MAX),
  logging + returning false; added <climits>.
- socket.cpp: replace the non-atomic _socket_initialized check-then-act WSAStartup
  init with a std::once_flag + std::call_once helper (initialize_winsock); added
  <mutex>. Fixes the race when Sockets are constructed concurrently.
- rename socket_msvc.hpp -> socket_win32.hpp (guard is _WIN32, applies to any
  Windows toolchain) and update all #include references (socket, rtsp, lib umbrella).
