+++ components/rtsp/include/rtsp_session.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "socket_msvc.hpp"
+#include "socket_win32.hpp"
 
 #include <functional>
 #include <memory>
