+++ components/socket/include/tcp_socket.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "socket_msvc.hpp"
+#include "socket_win32.hpp"
 
 #ifndef _WIN32
 #include <netinet/tcp.h>
